Replace monolithic fibermorph.py with compatibility shim, complete refactoring
